refactor(launch-lib): split into 9 launch-lib/ submodules
Brings packages/crane-mcp/src/cli/launch-lib.ts under structural caps
(was 2308 LOC, the largest violator in the portfolio).

- launch-lib.ts (94 LOC): pure barrel re-export + thin main() delegating
  to runMain().
- launch-lib/constants.ts: shared constants, ENGAGEMENT_REGISTRY,
  INFISICAL_PATHS, VentureWithRepo, EngagementContext. CRANE_CONSOLE_ROOT
  resolves 5 levels up (one deeper than original — compiled output moved
  one directory).
- launch-lib/secrets.ts: ensureInfisicalConfig, fetchSecrets (split into
  resolveInfisicalTarget, checkSpawnResult, parseInfisicalOutput).
- launch-lib/build-utils.ts: ensureFreshBuild, syncVentureRepo (extracted
  makeGitOut, readRepoSyncState, evaluateSyncState, applyFastForward).
- launch-lib/skill-sync.ts: syncClaudeAssets, syncGlobalSkills,
  syncVentureSkills, parseSkillScope, loadClaudeDenyRules.
- launch-lib/mcp-setup.ts: MCP setup; ensureParallelIsolationHooks split
  into installParallelIsolationScripts, buildParallelIsolationHookEntries,
  mergeParallelIsolationHooks.
- launch-lib/engagement.ts: parseEngagementArg, listClientEngagements.
- launch-lib/venture.ts: fetchVentures, matchVenturesToRepos, cloneVenture,
  printVentureList, promptSelection.
- launch-lib/agent-launch.ts: resolveAgent, validateAgentBinary,
  stripAgentFlags, extractPassthroughArgs, getStartupPrompt, launchAgent,
  assertEngagementScope, launchEngagement. buildChildEnv refactored to
  use VentureIdentity object (was 6 params).
- launch-lib/main-router.ts: routeDirectLaunch, handleSecretsAudit,
  printHelp, runMain (keeps main() under 10 lines).

Strict dependency DAG, no cycles. Public API preserved via re-exports.
764/764 tests pass.
